Move VirtualDrive to conftest so it can be used in other tests as well
